+++ src/dpm_feature.cpp
@@ -85,7 +85,7 @@ void Feature::computeFeaturePyramid(const Mat &imageM, vector< Mat > &pyramid)
 #endif
     for (int i = 0; i < params.interval; i++)
     {
-        const double scale = (double)(1.0f/pow(params.sfactor, i));
+        const double scale = (double)(1.0f/pow(params.sfactor, i)); // cannot modify !
         Mat imScaled;
         resize(imageM, imScaled, imSize * scale);
         // First octave at twice the image resolution
@@ -184,15 +184,15 @@ void Feature::computeHOG32D(const Mat &imageM, Mat &featM, const int sbin, const
     CV_Assert(pad_x >= 0);
     CV_Assert(pad_y >= 0);
     CV_Assert(imageM.channels() == 3);
-    CV_Assert(imageM.depth() == CV_64F);
+    CV_Assert(imageM.depth() == CV_32F);
 
     // epsilon to avoid division by zero
-    const double eps = 0.0001;
+    const float eps = 0.0001;
     // number of orientations
     const int numOrient = 18;
     // unit vectors to compute gradient orientation
-    const double uu[9] = {1.000, 0.9397, 0.7660, 0.5000, 0.1736, -0.1736, -0.5000, -0.7660, -0.9397};
-    const double vv[9] = {0.000, 0.3420, 0.6428, 0.8660, 0.9848,  0.9848,  0.8660,  0.6428,  0.3420};
+    const float uu[9] = {1.000, 0.9397, 0.7660, 0.5000, 0.1736, -0.1736, -0.5000, -0.7660, -0.9397};
+    const float vv[9] = {0.000, 0.3420, 0.6428, 0.8660, 0.9848,  0.9848,  0.8660,  0.6428,  0.3420};
 
     // image size
     const Size imageSize = imageM.size();
@@ -208,9 +208,9 @@ void Feature::computeHOG32D(const Mat &imageM, Mat &featM, const int sbin, const
     const Size visible = blockSize*sbin;
 
     // initialize historgram, norm, output feature matrices
-    Mat histM = Mat::zeros(Size(blockSize.width*numOrient, blockSize.height), CV_64F);
-    Mat normM = Mat::zeros(Size(blockSize.width, blockSize.height), CV_64F);
-    featM = Mat::zeros(Size(outSize.width*dimHOG, outSize.height), CV_64F);
+    Mat histM = Mat::zeros(Size(blockSize.width*numOrient, blockSize.height), CV_32F);
+    Mat normM = Mat::zeros(Size(blockSize.width, blockSize.height), CV_32F);
+    featM = Mat::zeros(Size(outSize.width*dimHOG, outSize.height), CV_32F);
 
     // get the stride of each matrix
     const size_t imStride = imageM.step1();
@@ -219,45 +219,45 @@ void Feature::computeHOG32D(const Mat &imageM, Mat &featM, const int sbin, const
     const size_t featStride = featM.step1();
 
     // calculate the zero offset
-    const double* im = imageM.ptr<double>(0);
-    double* const hist = histM.ptr<double>(0);
-    double* const norm = normM.ptr<double>(0);
-    double* const feat = featM.ptr<double>(0);
+    const float* im = imageM.ptr<float>(0);
+    float* const hist = histM.ptr<float>(0);
+    float* const norm = normM.ptr<float>(0);
+    float* const feat = featM.ptr<float>(0);
 
     for (int y = 1; y < visible.height - 1; y++)
     {
         for (int x = 1; x < visible.width - 1; x++)
         {
             // OpenCV uses an interleaved format: BGR-BGR-BGR
-            const double* s = im + 3*min(x, imageM.cols-2) + min(y, imageM.rows-2)*imStride;
+            const float* s = im + 3*min(x, imageM.cols-2) + min(y, imageM.rows-2)*imStride;
 
             // blue image channel
-            double dyb = *(s+imStride) - *(s-imStride);
-            double dxb = *(s+3) - *(s-3);
-            double vb = dxb*dxb + dyb*dyb;
+            float dyb = *(s+imStride) - *(s-imStride);
+            float dxb = *(s+3) - *(s-3);
+            float vb = dxb*dxb + dyb*dyb;
 
             // green image channel
             s += 1;
-            double dyg = *(s+imStride) - *(s-imStride);
-            double dxg = *(s+3) - *(s-3);
-            double vg = dxg*dxg + dyg*dyg;
+            float dyg = *(s+imStride) - *(s-imStride);
+            float dxg = *(s+3) - *(s-3);
+            float vg = dxg*dxg + dyg*dyg;
 
             // red image channel
             s += 1;
-            double dy = *(s+imStride) - *(s-imStride);
-            double dx = *(s+3) - *(s-3);
-            double v = dx*dx + dy*dy;
+            float dy = *(s+imStride) - *(s-imStride);
+            float dx = *(s+3) - *(s-3);
+            float v = dx*dx + dy*dy;
 
             // pick the channel with the strongest gradient
             if (vg > v) { v = vg; dx = dxg; dy = dyg; }
             if (vb > v) { v = vb; dx = dxb; dy = dyb; }
 
             // snap to one of the 18 orientations
-            double best_dot = 0;
+            float best_dot = 0;
             int best_o = 0;
             for (int o = 0; o < (int)numOrient/2; o++)
             {
-                double dot =  uu[o]*dx + vv[o]*dy;
+            	float dot =  uu[o]*dx + vv[o]*dy;
                 if (dot > best_dot)
                 {
                     best_dot = dot;
@@ -271,14 +271,14 @@ void Feature::computeHOG32D(const Mat &imageM, Mat &featM, const int sbin, const
             }
 
             // add to 4 historgrams around pixel using bilinear interpolation
-            double yp =  ((double)y+0.5)/(double)sbin - 0.5;
-            double xp =  ((double)x+0.5)/(double)sbin - 0.5;
+            float yp =  ((float)y+0.5)/(float)sbin - 0.5;
+            float xp =  ((float)x+0.5)/(float)sbin - 0.5;
             int iyp = (int)floor(yp);
             int ixp = (int)floor(xp);
-            double vy0 = yp - iyp;
-            double vx0 = xp - ixp;
-            double vy1 = 1.0 - vy0;
-            double vx1 = 1.0 - vx0;
+            float vy0 = yp - iyp;
+            float vx0 = xp - ixp;
+            float vy1 = 1.0 - vy0;
+            float vx1 = 1.0 - vx0;
             v = sqrt(v);
 
             // fill the value into the 4 neighborhood cells
@@ -300,9 +300,9 @@ void Feature::computeHOG32D(const Mat &imageM, Mat &featM, const int sbin, const
     // compute the energy in each block by summing over orientation
     for (int y = 0; y < blockSize.height; y++)
     {
-        const double* src = hist + y*histStride;
-        double* dst = norm + y*normStride;
-        double const* const dst_end = dst + blockSize.width;
+        const float* src = hist + y*histStride;
+        float* dst = norm + y*normStride;
+        float const* const dst_end = dst + blockSize.width;
         // for each cell
         while (dst < dst_end)
         {
@@ -323,9 +323,9 @@ void Feature::computeHOG32D(const Mat &imageM, Mat &featM, const int sbin, const
     {
         for (int x = pad_x; x < outSize.width - pad_x; x++)
         {
-            double* dst = feat + y*featStride + x*dimHOG;
-            double* p, n1, n2, n3, n4;
-            const double* src;
+        	float *dst = feat + y*featStride + x*dimHOG;
+        	float *p, n1, n2, n3, n4;
+            const float* src;
 
             p = norm + (y - pad_y + 1)*normStride + (x - pad_x + 1);
             n1 = 1.0f / sqrt(*p + *(p + 1) + *(p + normStride) + *(p + normStride + 1) + eps);
@@ -336,17 +336,17 @@ void Feature::computeHOG32D(const Mat &imageM, Mat &featM, const int sbin, const
             p = norm + (y - pad_y)*normStride + x - pad_x;
             n4 = 1.0f / sqrt(*p + *(p + 1) + *(p + normStride) + *(p + normStride + 1) + eps);
 
-            double t1 = 0.0, t2 = 0.0, t3 = 0.0, t4 = 0.0;
+            float t1 = 0.0, t2 = 0.0, t3 = 0.0, t4 = 0.0;
 
             // contrast-sesitive features
             src = hist + (y - pad_y + 1)*histStride + (x - pad_x + 1)*numOrient;
             for (int o = 0; o < numOrient; o++)
             {
-                double val = *src;
-                double h1 = min(val*n1, 0.2);
-                double h2 = min(val*n2, 0.2);
-                double h3 = min(val*n3, 0.2);
-                double h4 = min(val*n4, 0.2);
+            	float val = *src;
+            	float h1 = min(val*n1, (float)0.2);
+            	float h2 = min(val*n2, (float)0.2);
+            	float h3 = min(val*n3, (float)0.2);
+            	float h4 = min(val*n4, (float)0.2);
                 *(dst++) = 0.5 * (h1 + h2 + h3 + h4);
                 src++;
                 t1 += h1;
@@ -359,11 +359,11 @@ void Feature::computeHOG32D(const Mat &imageM, Mat &featM, const int sbin, const
             src =  hist + (y - pad_y + 1)*histStride + (x - pad_x + 1)*numOrient;
             for (int o = 0; o < numOrient/2; o++)
             {
-                double sum = *src + *(src + numOrient/2);
-                double h1 = min(sum * n1, 0.2);
-                double h2 = min(sum * n2, 0.2);
-                double h3 = min(sum * n3, 0.2);
-                double h4 = min(sum * n4, 0.2);
+            	float sum = *src + *(src + numOrient/2);
+            	float h1 = min(sum * n1, (float)0.2);
+            	float h2 = min(sum * n2, (float)0.2);
+            	float h3 = min(sum * n3, (float)0.2);
+            	float h4 = min(sum * n4, (float)0.2);
                 *(dst++) = 0.5 * (h1 + h2 + h3 + h4);
                 src++;
             }
@@ -387,7 +387,7 @@ void Feature::computeHOG32D(const Mat &imageM, Mat &featM, const int sbin, const
             if (m > pad_y - 1 && m < featM.rows - pad_y && n > pad_x*dimHOG - 1 && n < featM.cols - pad_x*dimHOG)
                 continue;
 
-            featM.at<double>(m, n + dimHOG - 1) = 1;
+            featM.at<float>(m, n + dimHOG - 1) = 1;
         } // for x
     }// for y
 }
@@ -407,10 +407,10 @@ void Feature::projectFeaturePyramid(const Mat &pcaCoeff, const std::vector< Mat 
         int width = orgM.cols/dimHOG;
         int height = orgM.rows;
         // initialize the project feature matrix
-        Mat projM = Mat::zeros(height, width*dimPCA, CV_64F);
+        Mat projM = Mat::zeros(height, width*dimPCA, CV_32F);
         //get the pointer of the matrix
-        double* const featOrg = orgM.ptr<double>(0);
-        double* const featProj = projM.ptr<double>(0);
+        float* const featOrg = orgM.ptr<float>(0);
+        float* const featProj = projM.ptr<float>(0);
 
         // get the stride of each matrix
         const size_t orgStride = orgM.step1();
@@ -420,15 +420,15 @@ void Feature::projectFeaturePyramid(const Mat &pcaCoeff, const std::vector< Mat 
         {
             for (int x = 0; x < width; x++)
             {
-                double* proj = featProj + y*projStride + x*dimPCA;
+            	float* proj = featProj + y*projStride + x*dimPCA;
                 // for each pca dimension
                 for (int c = 0; c < dimPCA; c++)
                 {
-                    double* org = featOrg + y*orgStride + x*dimHOG;
+                	float* org = featOrg + y*orgStride + x*dimHOG;
                     // dot product 32d HOG feature with the coefficient vector
                     for (int r = 0; r < dimHOG; r++)
                     {
-                        *proj += *org * pcaCoeff.at<double>(r, c);
+                        *proj += *org * pcaCoeff.at<float>(r, c);
                         org++;
                     }
                     proj++;
@@ -442,25 +442,25 @@ void Feature::projectFeaturePyramid(const Mat &pcaCoeff, const std::vector< Mat 
 
 void Feature::computeLocationFeatures(const int numLevels, Mat &locFeature)
 {
-    locFeature = Mat::zeros(Size(numLevels, 3), CV_64F);
+    locFeature = Mat::zeros(Size(numLevels, 3), CV_32F);
 
     int b = 0;
     int e = min(numLevels, params.interval);
 
     for (int x = b; x < e; x++)
-        locFeature.at<double>(0, x) = 1;
+        locFeature.at<float>(0, x) = 1;
 
     b = e;
     e = min(numLevels, 2*e);
 
     for (int x = b; x < e; x++)
-        locFeature.at<double>(1, x) = 1;
+        locFeature.at<float>(1, x) = 1;
 
     b = e;
     e = min(numLevels, 3*e);
 
     for (int x = b; x < e; x++)
-        locFeature.at<double>(2, x) = 1;
+        locFeature.at<float>(2, x) = 1;
 }
 } // namespace dpm
 } // namespace cv
